Add no_restart property
- Add an option 'no_restart' (default to false for compatibility).
This allows to prevent service restart on configuration change.
The restart is often unwanted as it empties the cache.
The restart can be unnecessary (example: if we fix a comment...)
The restart can be delayed and orchestrated by an external tool, very
useful to deal with distributed environment (we probably do not want to
restart all nodes at the same time).

Signed-off-by: Cyril Martin <c.martin@criteo.com>