More control of (package) installation
- commit b2e98e6 change how the user was created. It said:
"If the user gets removed later on we should be recreating it".
But it caused an issue for us, because the user is already created by
the package and because we cannot update the HOME of a user executing a
program (and obviously we do not want to restart memcached for that...)
- By the way, we do not want to execute _package recipe, because Criteo
env requires to use the yum option --enablerepo=criteo-build and because
we consider to switch on a tarball installation (compiled by our-self).

Signed-off-by: Cyril Martin <c.martin@criteo.com>